Double sebesség értékek lehetővé tétele

+++ src/main/java/model/board/element/Empty.java
@@ -3,7 +3,7 @@ package model.board.element;
 import java.awt.*;
 
 public class Empty extends Entity{
-    public Empty(int x, int y, int width, int height) {
+    public Empty(double x, double y, int width, int height) {
         super(x, y, width, height, 0, null, false, false);
     }
