+++ src/main/java/model/board/element/character/BasicMonster.java
@@ -36,7 +36,7 @@ public class BasicMonster extends Monster {
      * @param visible  the status indicating if the monster is visible
      * @param board    the game board the monster belongs to
      */
-    public BasicMonster(int x, int y, int width, int height, double velocity, List<Image> images, boolean alive, boolean visible, Board board) {
+    public BasicMonster(double x, double y, int width, int height, double velocity, List<Image> images, boolean alive, boolean visible, Board board) {
         super(x, y, width, height, velocity, images, alive, visible, board);
     }
