+++ src/main/java/model/board/element/character/IntelligentMonster.java
@@ -32,7 +32,7 @@ public class IntelligentMonster extends Monster {
      * @param visible  the status indicating if the monster is visible
      * @param board    the game board the monster belongs to
      */
-    public IntelligentMonster(int x, int y, int width, int height, double velocity, List<Image> images, boolean alive, boolean visible, Board board) {
+    public IntelligentMonster(double x, double y, int width, int height, double velocity, List<Image> images, boolean alive, boolean visible, Board board) {
         super(x, y, width, height, velocity, images, alive, visible, board);
     }
