+++ src/main/java/model/board/element/powerup/benefit/DetonatorBonus.java
@@ -6,7 +6,7 @@ import model.board.element.powerup.Bonus;
 import java.awt.*;
 
 public class DetonatorBonus extends Bonus {
-    public DetonatorBonus(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Player owner) {
+    public DetonatorBonus(double x, double y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Player owner) {
         super(x, y, width, height, velocity, image, alive, visible, owner);
     }
